# Use static data ReadOnlySpan<byte> optimize GeneratedId