Use static data ReadOnlySpan<byte> optimize GeneratedId, gain a little performance improve and eliminate initial char array in the heap. 
https://vcsjones.dev/2019/02/01/csharp-readonly-span-bytes-static/

Summary of the changes (Less than 80 chars)
 - Use `ReadOnlySpan<byte>` avoid static char array allocation in the heap.  
 - Improve performance a little. 

I have created a benchmark:  https://gist.github.com/Ruikuan/1066fb7af4deb0f6d50319ef1d98717b , result:

```
BenchmarkDotNet=v0.11.5, OS=Windows 10.0.18362
Intel Core i7-3720QM CPU 2.60GHz (Ivy Bridge), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100
  [Host]     : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), 64bit RyuJIT
  DefaultJob : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), 64bit RyuJIT
```

|            Method |     Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------ |---------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| GenerateId_Origin | 18.19 ns | 0.1217 ns | 0.1079 ns |  1.00 | 0.0153 |     - |     - |      48 B |
|        GenerateId | 16.50 ns | 0.0786 ns | 0.0697 ns |  0.91 | 0.0153 |     - |     - |      48 B |


Other:
* `GenerateId` is `unsafe` again.
* I don't have a big endian computer, so I can't test the big endian part myself.