@khellang 2019-10-08
I would say this optimization isn't worth the added code (complexity) for a 2ns improvement (which could very well be noise) 🤷‍♂