@Ruikuan 2019-10-08
@khellang I have run this benchmark many times, the new one is faster every time, although mostly only about 0.95. Above is the lastest result. 

Yes, the code is more complex. If we have utf8 string now, it would be better.